+++ src/ui/switchboard_listener/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from .listener_tab import SwitchboardListenerTab
+
+__all__ = ["SwitchboardListenerTab"]
+
+